requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
openai>=1.1.1
python-dotenv>=1.0.0
fastapi>=0.95.2
uvicorn>=0.22.0
pydantic>=2.0.3
tiktoken>=0.4.0
unstructured>=0.10.30
python-multipart>=0.0.6
jinja2>=3.1.2 